test: multi-backend loop-factory residual in dedicated module
Move the #1509 review residual into its own test module so ruff format
does not thrash the large existing pytester file. Covers asyncio +
analogous non-asyncio backend marks with loop-type asserts (#1463 shape).
